Training after the debugging
Uncommenting two lines in the deform_volume fucntion made the dice coefficient scores better. I tested with lesser epochs from the server and it worked well. But I could not push the changes from the server since there were changes on my local(windows) machine as well. And the jobs that I tried to submit on the 20th Dec, 2024 were all put on hold with the message that the machines did not match my requirements. I have to figure this out.